+++ api/src/main/java/org/allaymc/api/block/type/BlockStateSafeGetter.java
@@ -1,8 +1,11 @@
 package org.allaymc.api.block.type;
 
 import lombok.experimental.UtilityClass;
+import lombok.extern.slf4j.Slf4j;
 import org.allaymc.api.block.property.type.BlockPropertyType;
+import org.allaymc.api.network.ProtocolInfo;
 import org.allaymc.api.registry.Registries;
+import org.allaymc.api.utils.HashUtils;
 import org.allaymc.api.utils.Identifier;
 import org.allaymc.updater.block.BlockStateUpdaters;
 import org.cloudburstmc.nbt.NbtMap;
@@ -10,16 +13,60 @@ import org.cloudburstmc.nbt.NbtMapBuilder;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.TreeMap;
 
 /**
  * BlockStateSafeGetter is used to get a block state safely, which means that
  * the plugin developers can get an item type without worrying about being broken
  * in the next minecraft version compared to using {@link BlockTypes} directly.
  *
- * @author daoge_cmd
+ * @author daoge_cmd | harry-xi
  */
+@Slf4j
 @UtilityClass
 public final class BlockStateSafeGetter {
+
+    /**
+     * Get the block state from the block state NBT.
+     *
+     * @param nbt the block state NBT.
+     *
+     * @return the block state, or the default state of {@code BlockTypes.UNKNOWN}
+     * if the block state is not found, or the block state version is too new.
+     */
+    public static BlockState fromNBT(NbtMap nbt) {
+        var version = nbt.getInt("version");
+        if (version > ProtocolInfo.BLOCK_STATE_VERSION_NUM) {
+            log.warn("Block state version is too new: {}", nbt);
+            return BlockTypes.UNKNOWN.getDefaultState();
+        }
+
+        if (version < ProtocolInfo.BLOCK_STATE_VERSION_NUM) {
+            nbt = BlockStateUpdaters.updateBlockState(nbt, BlockStateUpdaters.LATEST_VERSION);
+        }
+
+        // Make sure that tree map is used
+        // If the map inside states nbt is not tree map
+        // the block state hash will be wrong!
+        var states = new TreeMap<>(nbt.getCompound("states"));
+        // To calculate the hash of the block state
+        // "name" field must be in the first place
+        var tag = NbtMap.builder()
+                .putString("name", nbt.getString("name"))
+                .putCompound("states", NbtMap.fromMap(states))
+                .build();
+        int blockStateHash = HashUtils.fnv1a_32_nbt(tag);
+
+        // Get block state by hash
+        BlockState blockState = Registries.BLOCK_STATE_PALETTE.get(blockStateHash);
+        if (blockState != null) {
+            return blockState;
+        }
+
+        log.warn("Block state not found: {}", nbt);
+        return BlockTypes.UNKNOWN.getDefaultState();
+    }
+
     /**
      * Get a block type by its name.
      *