+++ api/src/main/java/org/allaymc/api/utils/StructureException.java
@@ -1,0 +1,7 @@
+package org.allaymc.api.utils;
+
+import lombok.experimental.StandardException;
+
+@StandardException
+public class StructureException extends RuntimeException {
+}